opencvlib/requirements.txt
more_itertools==4.1.0
scikit_image==0.13.1
SQLAlchemy==1.2.3
scipy==1.0.0
pandas==0.22.0
matplotlib==2.1.2
joblib==0.11
numpy==1.14.1+mkl
fuckit==4.8.0
Pillow==5.0.0
beautifulsoup4==4.6.0
skimage==0.0
scikit_learn==0.19.1
sympy==1.1.1
tensorflow==1.6.0